Rework exception handling
We are now accounting for request timeouts, connection errors, json decoding errors, and more. When they occur, we retry every n seconds, with n increasing exponentially until we reach the wait time cap or we hit the maximum number of retries. In addition, we display a user-friendly error message instead of the full traceback.